+++ swap/src/asb/network.rs
@@ -42,11 +42,15 @@ pub mod transport {
         maybe_tor_client: Option<Arc<TorClient<TokioRustlsRuntime>>>,
         register_hidden_service: bool,
         num_intro_points: u8,
+        max_concurrent_rend_requests: usize,
     ) -> Result<OnionTransportWithAddresses> {
-        // Streams are multiplexed via yamux, we don't need more than one.
-        const MAX_STREAMS_PER_CIRCUIT: u32 = 2;
-        // Default value
-        const POW_QUEUE_DEPTH: usize = 8192;
+        // Streams are multiplexed via yamux, we don't really need more than one.
+        const MAX_STREAMS_PER_CIRCUIT: u32 = 4;
+        // This does not affect the PoW directly (only very slightly) but only serves as a protection
+        // against memory exhaustion attacks when the queue of intro request fills up
+        // We therefore set it to a fairly high value because there is barely any harm in doing so.
+        // `MAX_CONCURRENT_REND_REQUESTS` is much more important in terms of DOS protection.
+        const POW_QUEUE_DEPTH: usize = 2048;
 
         let (maybe_tor_transport, onion_addresses) = if let Some(tor_client) = maybe_tor_client {
             let mut tor_transport =
@@ -67,7 +71,7 @@ pub mod transport {
                     .build()
                     .expect("We specified a valid nickname");
 
-                match tor_transport.add_onion_service(onion_service_config, ASB_ONION_SERVICE_PORT)
+                match tor_transport.add_onion_service(onion_service_config, ASB_ONION_SERVICE_PORT, max_concurrent_rend_requests)
                 {
                     Ok(addr) => {
                         tracing::debug!(
@@ -102,7 +106,7 @@ pub mod transport {
 }
 
 pub mod behaviour {
-    use libp2p::{identify, identity, ping, swarm::behaviour::toggle::Toggle};
+    use libp2p::{connection_limits, identify, identity, ping, swarm::behaviour::toggle::Toggle};
     use swap_p2p::{out_event::alice::OutEvent, patches};
 
     use super::*;
@@ -115,6 +119,7 @@ pub mod behaviour {
     where
         LR: LatestRate + Send + 'static,
     {
+        connection_limits: connection_limits::Behaviour,
         pub rendezvous: Toggle<rendezvous::register::Behaviour>,
         pub quote: quote::Behaviour,
         pub swap_setup: alice::Behaviour<LR>,
@@ -141,6 +146,7 @@ pub mod behaviour {
             env_config: env::Config,
             identify_params: (identity::Keypair, XmrBtcNamespace),
             rendezvous_nodes: Vec<PeerId>,
+            connection_limits: connection_limits::ConnectionLimits,
         ) -> Self {
             let (identity, namespace) = identify_params;
             let agent_version = format!("asb/{} ({})", env!("CARGO_PKG_VERSION"), namespace);
@@ -162,6 +168,7 @@ pub mod behaviour {
             };
 
             Self {
+                connection_limits: connection_limits::Behaviour::new(connection_limits),
                 rendezvous: Toggle::from(behaviour),
                 quote: quote::alice(),
                 swap_setup: alice::Behaviour::new(